+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - Added sample support for [TemplatedConfigLoader](https://kedro.readthedocs.io/en/latest/kedro.config.TemplatedConfigLoader.html)
+- MLFlow support updated to not use nested runs.
 
 ## [0.1.9] - 2021-01-08